Mutation Tester Cleanup and Compilation Validation
Mutation Tester now has a method to run an Ant Target specific to the
file system of this project, and methods which take Test and Mutant
objects and use these to produce the correct cmd line ant task.

Code is now checked to see if it compiles when it is submitted; if it
does not, it is rejected and they must try again.

Multiple Declarations of Game State and Mutation Tester have been
recombined, so now Attacker Page and Defender Page hold a copy of each
as a variable to refer to.

Build File has been modified to clean up poor dependencies which would
have caused code to be recompiled many times. Compilation has been
separated into two different kinds, compile-mutant and compile-test.