More Game Manager Logic, Platform Compatibility Fix
Game Manager Logic for creating mutants and tests has been implemented,
and mutants can now be added to the database through basic gameplay.

Platform compatibility should be fixed. If the properties file for a
variable exists it will be used (E.g. on Mac and Linux, using the
BuildFile), otherwise it will default to my versions, as mine should be
the only Windows system this will need to operate on at present.