Target Executions, Timestamps, MutTester Update
Added target executions to track results of compiles and test runs.
Added timestamps to keep track of chronological order of tests/mutants
and to gauge time taken as a partial measure of complexity. Updating the
mutation tester to provide more information, partially through the use
of the new Target Executions.