Fix wrong redirect in admin page
To keep the style layout, all tabs inherit the <a> style sheet.
Non-active tabs link to their sides using href="...=".
The active tab however does not need to link anywhere. It previously did
link to "#", which results in a redirect to "/#", which is the start
page. The active tab now doesn't link anywhere, but still looks like a
tab.