Update DatePicker in Create Battleground
Fixed the representation of start and finish dates as well as the hours and minutes, when first loading the page and after selecting new dates to YY/MM/DD (ISO).
Adjusted the update of changed dates, so that they stay selected in the calendar and kept for submitting the form.
Added checks and error messages for input fields, in case the user fills in an invalid date and/or time; e.g. a date in the past (which is also handled by setting the date back to today).