Update contact us page
The feedback html form is now only visible when `sending emails` is enabled.
Previously it was checked whether the email address STMP host was not empty, which didn't check for validity or setup completeness. In admin system settings a complete email configuration is enforced (validity is currently not checked yet), depending on the `sending emails` enabled, so checking for `sending emails` here is a good first step.