Added Validation to date format in Create Battleground
Added JavaScript function handling the validation of date format, which must now be in format YYYY/MM/DD and containing valid input (e.g. maximum number of days in a month, taking into account leap years). A warning message appears and creation of a battleground is disabled, when the user enters an invalid date.