Update UploadManager to zip files for test/mutant
Instead of multiple .java files, the UploadManager servlet now accepts
upload of multiple mutants and tests through a zip file for each mutants
and tests.

Refactor parts of the UploadManager to ease implementation of
dependencies in the future.

Updated upload_class_view jsp to match zip upload.

Update ZipFileUtils to use ArrayList instead of LinkedList, as the list
is accessed using the index.