Enable dependency upload for mutants/tests.
Add calls to store dependencies to database.
Add calls to remove dependencies from database once request failed.
Restructured some code to collect dependencies to add them to the database. Added internal class for this, too.
Extended addition of mutants and tests to allow compilation with dependencies.

Extended class upload view to display how many dependencies were uploaded. This is just a draft and shouldn't be the in the next release.