Fix mutant compilation with dependencies
Submitted mutants could not be compiled, because dependencies were missing. Once the directory of the class
under test was linked, the mutants could be compiled, but tests couldn't kill mutants, because the CUT was
in the classpath too and overwrote the mutant.

These changes put all dependencies (`.java` and `.class` files) of a class under test in a separate
`dependencies/` folder, similar to `mutants/` and `tests/` for uploaded mutants and tests.
UploadManager enforces these folder changes.
Compiler moves generated .class files to the dependencies folder after successful compilation.

Update AntRunner do use the CUT dependency directory as an additional classpath entry for execution of ant
target. *No changes to the ant target configurations had to be made*, because a `cuts.deps` configuration
existed, but was unused before (since 2016 lmfao). This allows existing Code Defender front-end docker containers
to not be required to be rebuilt, neat.