Fix mutant view for games with uploaded mutants and tests
When uploading a class with mutants and tests, they are executed against
each other. This results in entries to the `targetexecutions` table.
When creating a game with mutants and tests, both are copied, not
executed against each other. This greatly reduces the time it takes to
start such a game.

#445 added the kill messages by looking at the `targetexecutions` table
to finds out which test killed which mutant and therefore find the kill
message. However, atm we don't have an entry in the `targetexecutions`
table for copied mutants and tests. The queried test is then null and
the whole `mutants_list.jsp` fails.