Copy target execution what starting game
Instead of just the mutants and tests, the target executions are now copied, too.

Updated how copied mutants are killed.
Previously, the killmap was used and the first successful test killed a mutant.
Now, the actual test which originally killed the mutant kills the copied mutant.

Also, the original kill message is now copied, too.