Add ConfigFileResolvers and multiple Configuration implementations
The ConfigFileResolver (CFR) is a abstraction over the process of getting
a Reader for accessing the content of a configuration file.
We currently have six of those, which try to load the file from different
locations.
- SystemPropertyCFR: Load file from location supplied by system property
- EnvironmentVariableCFR: Load file from loc. supplied by EnvVar
- TomcatConfigCFR: Load file from $CATALINA_BASE/conf folder
- ClasspathCFR: Tries to find file on classpath
- MockedCFR: Returns a StringReader, so we can test the ConfigFile parsing
             without a file or the need to mock a file.

This also adds a more Configuration (C) classes more precisely:
- SystemPropertyC: Read values from system properties
- EnvironmentVariableC: Read values from environment variables
- PropertiesFileC: Read values from properties file using the CFRs
- TieredC: Tries to load the values from the three options mentiond above in
           order.

Currently the system works by specifing variables (with default values) in the
Configuration class, which the *Configuration classes iterate over.
Those classes change the format from lowerCamelCase to their prefered formats
which are system.property, ENVIRONMENT_VARIABLES and properties.file and then
try to lookup those names.