Remove code for handling default values in Configuration class
As we do not set default values in the Configuration class anymore, but through
the `src/resources/codedefenders.properties` file, we can get rid of all code
which ensured we only overwrite properties with user set properties and not
with the inherited default values.
This will only work as long as all attributes in the Configuration class are
initialized with null, so I added some (more) warnings to the documentation of
the Configuration class.