Fix exception while parsing the mutants json of the MutantAccordion
The description of the problematic mutants contained a `'`.

We first escaped this via `StringEscapeUtils::escapeJavaScript` inside the
MutantDTO constructor and then later HTML escaped it while converting it to a
JSON string with the `Gson::toJson` method.

This converted "'" to "\\\u0027" while it actually should have been converted
to "\\u0027".
I could not drop the `::escapeJavaScript` call, since then for example "\""
would not be escaped properly.

So with first converting `::toJson` and afterwards `::escapeJavaScript` it will
now properly escape all characters.