Fix Code 1093 error when runing the V9 migration script on MySQL 8
In MySQL 8 it is not allowed to update a table which is used in an inner
SELECT statement.
To work around this problem we wrap the inner SELECT in another SELECT and
assign an alias to the wrapping SELECT statement. MySQL will—because of the
alias—create a temporary table which can safely be used in the WHERE part of
the UPDATE statement.