Refactor maven build setup
The versions of all build plugins are now configured via the <pluginManagement>
element. This also includes versions of predefined build plugins.

The build also now enforces the usage of at least maven 3.6

The dependency plugin can be used, for example, to find unused or undeclared
dependencies.
The versions plugin can be used, for example, to list the maven plugins and
dependencies that can be updated.