Refactor CodeValidatorTest
Most test cases are now realized via JUnit 5 parameterized tests.

Data for the testcases is located in the resources directory and specific to
this test class.
One test input set (either one .txt containing a test or a directory containing
a original.txt and mutated.txt for a mutant) should only be used for exactly
one test case to make maintenance easier.