Include git.properties file in .war and introduce release profile
The git.properties file includes some information regarding the current state
of the git repository.
Those are mainly: The id, short id, and timestamp of the latest commit and
whether the repository is dirty or not.

The release profile enables deterministic builds and sets the timestamp of the
artifacts to be the same as the commit timestamp.
With the release profile disabled the builds are non-deterministic.
This should fix problems with tomcat not recompiling cached JSP files.

Additionally the release profile enforces the repository to not be in a dirty
state.
Note: New files (ignored or not ignored) do not make the repository 'dirty'

The release profile is enabled for all builds in the CI.