Fix container image build
We cache dependencies in the container image build by invoking
'mvn dependency:go-offline' rather early because those should not change much.

But javamelody depends on some strange/old artifacts that maven will not
resolve and instead fail.
Because those are not necessary for the later build we can simply ignore/mask
the failure with '|| true'.